[glyf] Split loca.hh
