+++ src/Makefile.sources
@@ -89,6 +89,7 @@ HB_BASE_sources = \
 	hb-ot-layout-gpos-table.hh \
 	hb-ot-layout-gsub-table.hh \
 	OT/glyf/glyf.hh \
+	OT/glyf/loca.hh \
 	OT/Layout/GSUB/Common.hh \
 	OT/Layout/GSUB/Sequence.hh \
 	OT/Layout/GSUB/SingleSubstFormat1.hh \
