+++ src/meson.build
@@ -93,6 +93,7 @@ hb_base_sources = files(
   'hb-ot-layout-gpos-table.hh',
   'hb-ot-layout-gsub-table.hh',
   'OT/glyf/glyf.hh',
+  'OT/glyf/loca.hh',
   'OT/Layout/GSUB/Common.hh',
   'OT/Layout/GSUB/Sequence.hh',
   'OT/Layout/GSUB/SingleSubstFormat1.hh',
